backend/requirements.txt
fastapi
uvicorn
python-multipart
pypandoc
pdf2docx 
python-pptx 
python-docx 
reportlab 
docx2pdf 
markdown